# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Lara Clinical Med",
        "version": "0.1"
    },
    "paths": {
...
        "/api/doctors": {
            "get": {
                "tags": [
                    "Médicos"
                ],
                "summary": "Listar médicos",
                "description": "Endpoint para listar médicos",
                "operationId": "4479c235c75c92e2425c15211470f478",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filtrar por nome do médico",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de médicos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sucesso"
                                        },
                                        "doctors": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Dra. Alessandra Moura"
                                                    },
                                                    "specialty": {
                                                        "type": "string",
                                                        "example": "Neurologista"
                                                    },
                                                    "cityId": {
                                                        "type": "integer",
                                                        "example": 1
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Houve um erro ao buscar os médicos."
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "Validation error message"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Médicos"
                ],
                "summary": "Criar um novo médico",
                "description": "Endpoint para criar um novo médico",
                "operationId": "ee9dd1ffdbcab8d6d47b2dd0bd06ed9b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Nome do médico",
                                        "type": "string",
                                        "example": "Dra. Alessandra Moura"
                                    },
                                    "specialty": {
                                        "description": "Especialidade do médico",
                                        "type": "string",
                                        "example": "Neurologista"
                                    },
                                    "cityId": {
                                        "description": "ID da cidade",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Médico criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Médico criado com sucesso."
                                        },
                                        "doctor": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Dra. Alessandra Moura"
                                                },
                                                "specialty": {
                                                    "type": "string",
                                                    "example": "Neurologista"
                                                },
                                                "cityId": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Não foi possível concluir seu cadastro."
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "Validation error message"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/doctors/consultation": {
            "post": {
                "tags": [
                    "Médicos"
                ],
                "summary": "Agendar uma consulta",
                "description": "Endpoint para agendar uma consulta com um médico",
                "operationId": "256079a8ec710cea977771e628580e4e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "doctorId": {
                                        "description": "ID do médico",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "patientId": {
                                        "description": "ID do paciente",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "date": {
                                        "description": "Data e hora da consulta",
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2023-10-01 10:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Consulta agendada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Consulta agendada com sucesso."
                                        },
                                        "consultation": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "doctorId": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "patient_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "date": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2023-10-01T10:00:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Não foi possível agendar a consulta."
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "Validation error message"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/doctors/{doctorId}/patients": {
            "get": {
                "tags": [
                    "Médicos"
                ],
                "summary": "Listar pacientes de um médico",
                "description": "Endpoint para listar pacientes de um médico",
                "operationId": "083703d5eb80db161ee83cbb7bcf3069",
                "parameters": [
                    {
                        "name": "doctorId",
                        "in": "path",
                        "description": "ID do médico",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "onlyScheduled",
                        "in": "query",
                        "description": "Consultas não realizadas",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filtrar por nome do paciente",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de pacientes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sucesso"
                                        },
                                        "patients": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "João da Silva"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "joao.silva@example.com"
                                                    },
                                                    "phone": {
                                                        "type": "string",
                                                        "example": "(11) 98765-4321"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Não foi possível listar os pacientes."
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "Validation error message"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}